backend/requirements.txt
fastapi
uvicorn
langchain
langchain-ollama
langchain-community
chromadb
openai
pydantic
pypdf
python-dotenv
